Revert "Revert "Fix protocolo""
This reverts commit 65d454f28b67de014c8c3249bcddb3916e75cb62.

+++ Codigo/Protocol.bas
@@ -630,8 +630,8 @@ Public Function HandleIncomingData(ByVal UserIndex As Integer, ByVal Message As 
 On Error Resume Next
     Set Reader = Message
     
-    Dim PacketId As Integer
-    PacketId = Reader.ReadInt16
+    Dim PacketId As Long
+    PacketId = Reader.ReadInt
     
 #If STRESSER = 1 Then
     Debug.Print "Paquete: " & PacketID
